requirements.txt
streamlit
openai-whisper
torch
reportlab
